fix: SJIP-1010 cleanup and use intl
