--- src/views/Analytics/Transcriptomic/ScatterPlotCanvas/index.module.css
@@ -1,27 +1,0 @@
-.tooltipContent {
-  display: flex;
-  gap: 8px;
-  flex-direction: column;
-  justify-content: flex-start;
-  text-align: left;
-}
-
-.tooltipText {
-  font-weight: 400;
-  font-size: 14px;
-}
-
-.tooltipLabel {
-  color: var(--gray-7);
-  margin-right: 4px;
-}
-
-.tooltipValue {
-  color: var(--gray-8);
-}
-
-.title {
-  color: var(--gray-9) !important;
-  font-size: 16px !important;
-  margin-bottom: 0 !important;
-}
