+++ src/views/Analytics/Transcriptomic/SwarmPlot/index.module.css
@@ -1,33 +1,4 @@
-.geneSymbol {
-  margin-left: 4px;
-  color: var(--gray-7);
-}
-
-.medianBoxT21 {
-  stroke: var(--blue-7);
-  fill: var(--blue-3);
-  stroke-width: 1;
-}
-
-.medianBoxControl {
-  stroke: var(--gray-6);
-  fill: var(--gray-4);
-  stroke-width: 1;
-}
-
-.limitLineT21 {
-  stroke: var(--blue-7);
-  stroke-width: 2;
-}
-
-.limitLineControl {
-  stroke: var(--gray-6);
-  stroke-width: 2;
-}
-
-
-.title {
-  color: var(--gray-9) !important;
-  font-size: 16px !important;
-  margin-bottom: 0 !important;
+.swarmPlot {
+    width: 100%;
+    height: 100%;
 }
