+++ src/views/Analytics/Transcriptomic/SwarmPlot/index.tsx
@@ -1,184 +1,159 @@
-import { useEffect, useMemo } from 'react';
 import intl from 'react-intl-universal';
-import { useSelector } from 'react-redux';
-import BasicDescription from '@ferlab/ui/core/components/BasicDescription';
-import SwarmPlotChart from '@ferlab/ui/core/components/Charts/SwarmPlot';
-import SwarmPlotBoxPlotSvgLayer from '@ferlab/ui/core/components/Charts/SwarmPlot/Layers/MedianBoxSvgLayer';
-import { ComputedDatum, SwarmRawDatum } from '@ferlab/ui/core/components/Charts/SwarmPlot/type';
-import { Typography } from 'antd';
+import Plot from 'react-plotly.js';
+import ChartSkeleton from '@ferlab/ui/core/components/Charts/Skeleton';
+import { Annotations, PlotData, PlotMouseEvent } from 'plotly.js';
 
-import { TTranscriptomicsDiffGeneExp } from 'services/api/transcriptomics/models';
-import { useTranscriptomicsSampleGeneExp } from 'store/analytics';
-import { transcriptomicsDiffGeneExpSelector } from 'store/analytics/selector';
+import { ITranscriptomicsSampleGeneExp } from '../../../../services/api/transcriptomics/models';
 
 import styles from './index.module.css';
 
-const { Title } = Typography;
-const CIRCLE_RADIUS = 9;
-
 export type TTranscriptomicsSwarmPlot = {
-  diffGeneExpId: string;
-  sampleIds: string[];
-  setSampleIdsCb: (sampleIds: string[]) => void;
+  selectedGeneSymbol: string;
+  selectedSampleIds: string[];
+  handleSampleSelection: (sampleIds: string[]) => void;
+  sampleGeneExp?: ITranscriptomicsSampleGeneExp;
+  loading: boolean;
 };
 
-export type TranscriptomicsSwarmRawDatum = SwarmRawDatum & {
-  group: string;
-  sample_id: string;
-  isSelected: boolean;
-};
-
-const getGeneSymbol = (diffGenes: TTranscriptomicsDiffGeneExp[], id: string) => {
-  let result = '';
-  diffGenes.forEach((diffGene) => {
-    diffGene.data.forEach((gene) => {
-      if (gene.ensembl_gene_id == id) {
-        result = gene.gene_symbol;
-        return;
-      }
-    });
-  });
-
-  return result;
-};
-
-const TranscriptomicsSwarmPlot = ({
-  diffGeneExpId,
-  setSampleIdsCb,
-  sampleIds,
+const SwarmPlotly = ({
+  selectedGeneSymbol,
+  handleSampleSelection,
+  selectedSampleIds,
+  sampleGeneExp,
+  loading,
 }: TTranscriptomicsSwarmPlot) => {
-  const diffGeneExp = useSelector(transcriptomicsDiffGeneExpSelector);
-  const sampleGeneExp = useTranscriptomicsSampleGeneExp(diffGeneExpId);
-  const t21GroupName = `T21 (${sampleGeneExp.data?.nT21})`;
-  const controlGroupName = `Control (${sampleGeneExp.data?.nControl})`;
-  const groups = [t21GroupName, controlGroupName];
-  const geneSymbol = useMemo(
-    () => getGeneSymbol(diffGeneExp.data ?? [], diffGeneExpId),
-    [diffGeneExp.data, diffGeneExpId],
-  );
+  const data = sampleGeneExp?.data ?? [];
 
-  const t21Group: TranscriptomicsSwarmRawDatum[] = [];
-  const controlGroup: TranscriptomicsSwarmRawDatum[] = [];
-  const data = sampleGeneExp?.data?.data ?? [];
+  const plotGroups = [
+    {
+      group: data.filter((e) => e.x === 1),
+      xValue: 1,
+      color: '#0284c7',
+      name: `${intl.get('screen.analytics.transcriptomic.swarmPlot.t21')} (${sampleGeneExp?.nT21})`,
+    },
+    {
+      group: data.filter((e) => e.x !== 1),
+      xValue: 2,
+      color: '#92a7c3',
+      name: `${intl.get('screen.analytics.transcriptomic.swarmPlot.control')} (${
+        sampleGeneExp?.nControl
+      })`,
+    },
+  ];
 
-  // To keep color scheme, data must be added by groups t21 and after control
-  data.forEach((node) => {
-    if (node.x == 1) {
-      t21Group.push({
-        ...node,
-        id: `t21-${node.sample_id}${Math.random()}`,
-        group: t21GroupName,
-        isSelected: node.sample_id == sampleIds[0],
-      });
-      return;
-    }
-    controlGroup.push({
-      ...node,
-      id: `control-${node.sample_id}${Math.random()}`,
-      group: controlGroupName,
-      isSelected: node.sample_id == sampleIds[0],
+  const plotData: PlotData[] = plotGroups.map(({ group, color, name, xValue }) => ({
+    y: group.map((e) => e.y),
+    x: Array(group.length).fill(xValue),
+    boxpoints: 'all',
+    name: name,
+    jitter: 1,
+    pointpos: 0,
+    type: 'box',
+    marker: {
+      color,
+      size: 7,
+      line: {
+        color: 'white',
+        width: 0.5,
+      },
+    },
+    text: group.map((e) =>
+      [
+        `${intl.get('screen.analytics.transcriptomic.swarmPlot.sample_id')} ${e.sample_id} <br>`,
+        `${intl.get('screen.analytics.transcriptomic.swarmPlot.fpkm')} ${e.y.toFixed(2)}`,
+      ].join(''),
+    ),
+    customdata: group.map((e) => e.sample_id),
+  })) as PlotData[];
+
+  const annotations: Partial<Annotations>[] = data
+    .filter((sample) => selectedSampleIds.includes(sample.sample_id))
+    .map((sample) => {
+      const xValue = sample.x === 1 ? 1 : 2;
+      return {
+        x: xValue,
+        y: sample.y,
+        arrowhead: 6,
+        ax: 100,
+        ay: 0,
+        bgcolor: 'rgba(255, 255, 255, 0.9)',
+        font: { size: 12 },
+        bordercolor: 'black',
+        borderwidth: 1,
+        text: `${sample.sample_id}: ${sample.y.toFixed(2)} ${intl.get(
+          'screen.analytics.transcriptomic.swarmPlot.fpkm',
+        )}`,
+      };
     });
-  });
 
-  const groupedData = [...t21Group, ...controlGroup];
+  if (loading) {
+    return <ChartSkeleton />;
+  }
 
-  useEffect(() => {
-    setSampleIdsCb([]);
-  }, [diffGeneExpId, setSampleIdsCb]);
+  const handlePlotClick = (data: Readonly<PlotMouseEvent>) => {
+    const clickedSample = (data.points.map((p) => p.customdata) as string[])[0];
+
+    if (selectedSampleIds.includes(clickedSample)) {
+      handleSampleSelection([]);
+    } else {
+      handleSampleSelection([clickedSample]);
+    }
+  };
 
   return (
-    <SwarmPlotChart
-      title={
-        <Title level={4} className={styles.title}>
-          {intl.get('screen.analytics.transcriptomic.swarmPlot.title')}
-          <span className={styles.geneSymbol}>{geneSymbol}</span>
-        </Title>
-      }
-      loading={sampleGeneExp.loading}
-      data={groupedData}
-      groups={groups}
-      spacing={8}
-      enableGridY={true}
-      enableGridX={false}
-      colors={['#0284c7', '#92a7c3']}
-      colorBy={'group'}
-      layers={[
-        'grid',
-        'axes',
-        ({ nodes }) => (
-          <SwarmPlotBoxPlotSvgLayer
-            nodes={nodes}
-            active={data.length > 0 && data.length === nodes.length}
-            groups={groups}
-            theme={{
-              [groups[0]]: {
-                rect: styles.medianBoxT21,
-                median: styles.medianBoxT21,
-                limit: styles.limitLineT21,
-              },
-              [groups[1]]: {
-                rect: styles.medianBoxControl,
-                median: styles.medianBoxControl,
-                limit: styles.limitLineControl,
-              },
-            }}
-          />
-        ),
-        'circles',
-        'annotations',
-        'mesh',
-      ]}
-      value={'y'}
-      size={CIRCLE_RADIUS}
-      margin={{ bottom: 32, left: 48, right: 0, top: 0 }}
-      onClick={(node: ComputedDatum<SwarmRawDatum>) => {
-        const sample = node as ComputedDatum<TranscriptomicsSwarmRawDatum>;
-        if (sample.data.sample_id === sampleIds[0]) {
-          setSampleIdsCb([]);
-          return;
-        }
-        setSampleIdsCb([sample.data.sample_id]);
-      }}
-      axisTop={null}
-      tooltip={(props) => {
-        const tooltipValue = props as ComputedDatum<TranscriptomicsSwarmRawDatum>;
-        return (
-          <div>
-            <BasicDescription
-              label={intl.get('screen.analytics.transcriptomic.swarmPlot.sample_id')}
-              text={tooltipValue.data.sample_id}
-            />
-            <BasicDescription
-              label={intl.get('screen.analytics.transcriptomic.swarmPlot.fpkm')}
-              text={`${tooltipValue.data.y.toFixed(2)}`}
-            />
-          </div>
-        );
-      }}
-      axisLeft={{
-        tickSize: 5,
-        tickPadding: 5,
-        tickRotation: 0,
-        legend: 'FPKM',
-        legendPosition: 'middle',
-        legendOffset: -32,
-      }}
-      annotations={[
-        {
-          type: 'circle',
-          match: {
-            data: {
-              sample_id: sampleIds[0] ?? -1,
-            },
+    <Plot
+      className={styles.swarmPlot}
+      data={plotData}
+      layout={{
+        annotations,
+        autosize: true,
+        title: {
+          text:
+            intl.get('screen.analytics.transcriptomic.swarmPlot.title') + ' ' + selectedGeneSymbol,
+          x: 0.05,
+          font: {
+            size: 16,
+            weight: 600,
           },
-          noteX: 30,
-          noteY: -30,
-          offset: 4,
-          note: `${sampleIds[0] ?? -1}`,
         },
-      ]}
+        margin: { l: 40, r: 10, t: 60, b: 40 },
+        legend: {
+          borderwidth: 1,
+          yanchor: 'top',
+          y: 0.99,
+          xanchor: 'right',
+          x: 0.99,
+        },
+        yaxis: {
+          title: intl.get('screen.analytics.transcriptomic.swarmPlot.fpkm'),
+        },
+        xaxis: {
+          tickvals: [1, 2],
+          ticktext: [
+            `${intl.get('screen.analytics.transcriptomic.swarmPlot.t21')} (${sampleGeneExp?.nT21})`,
+            `${intl.get('screen.analytics.transcriptomic.swarmPlot.control')} (${
+              sampleGeneExp?.nControl
+            })`,
+          ],
+        },
+      }}
+      useResizeHandler
+      config={{
+        modeBarButtonsToRemove: [
+          'toImage',
+          'resetGeo',
+          'lasso2d',
+          'sendDataToCloud',
+          'zoomIn2d',
+          'zoomOut2d',
+          'pan2d',
+          'select2d',
+        ],
+        displaylogo: false,
+      }}
+      onClick={handlePlotClick}
     />
   );
 };
 
-export default TranscriptomicsSwarmPlot;
+export default SwarmPlotly;
