--- src/views/Analytics/Transcriptomic/SwarmPlotly/index.module.css
@@ -1,4 +1,0 @@
-.swarmPlot {
-    width: 100%;
-    height: 100%;
-}
